requirements.txt
torch
torchvision
onnxruntime
numpy
pillow
fastapi
uvicorn
cerebrium